#182 relocate new files to target pkg
