+++ jdm-core/src/main/java/org/metricus/jdm/ui/AppIcon.java
@@ -1,4 +1,4 @@
-package jdiskmark;
+package org.metricus.jdm.ui;
 
 import java.awt.Color;
 import java.awt.Image;
