+++ jdm-core/src/test/java/jdiskmark/AppTest.java
@@ -3,6 +3,7 @@ package jdiskmark;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
+import org.metricus.jdm.ui.AppIcon;
 import static org.junit.jupiter.api.Assertions.*;
 
 /**
